+++ packages/out/make_A3_plot.py
@@ -21,7 +21,7 @@ def main(npd, cld_i, pd, clp):
     x_min, x_max, y_min, y_max = prep_plots.frame_max_min(
         cld_i['x'], cld_i['y'])
     asp_ratio = prep_plots.aspect_ratio(x_min, x_max, y_min, y_max)
-    coord, x_name, y_name = prep_plots.coord_syst(pd['coords'])
+    coord, x_name, y_name = "deg", "ra", "dec"
     x_zmin, x_zmax, y_zmin, y_zmax = prep_plots.frame_zoomed(
         x_min, x_max, y_min, y_max, clp['kde_cent'], clp['clust_rad'],
         pd['kp_ndim'], clp['KP_Bys_rt'])